Move bookcover to AddBookForm
