+++ src/components/AddBookForm.js
@@ -8,13 +8,17 @@ import placeholder from '../images/placeholder.png'
 export default function AddBookForm({
   onHandleSetSuccessMessage,
   onHandleCreateNewBook,
-  onGetBookCoverPreview,
   searchedBook,
 }) {
-  const [preview, setPreview] = useState(placeholder)
-  function getPreview(previewEvent) {
-    const preview = URL.createObjectURL(previewEvent.target.files[0])
-    setPreview(preview)
+  const [bookcover, setBookcover] = useState(placeholder)
+
+  function getBookcoverPreview(previewEvent) {
+    const bookcover = previewEvent.target.files[0]
+    const reader = new FileReader()
+    reader.onload = event => {
+      setBookcover(event.target.result)
+    }
+    reader.readAsDataURL(bookcover)
   }
 
   function handleSubmit(event) {
@@ -26,7 +30,7 @@ export default function AddBookForm({
       authors: authors.value,
       readingSince: readingSince.value ? readingSince.value : getToday(),
       onPage: onPage.value,
-      thumbnail: !searchedBook ? preview : searchedBook.thumbnail,
+      thumbnail: !searchedBook ? bookcover : searchedBook.thumbnail,
       isbn10: searchedBook ? searchedBook.isbn10 : '',
       isbn13: searchedBook ? searchedBook.isbn13 : '',
       year: searchedBook ? searchedBook.year : 'Unknown',
@@ -90,10 +94,10 @@ export default function AddBookForm({
             </div>
           ) : (
             <div id="bookcover-preview">
-              {preview === placeholder ? (
+              {bookcover === placeholder ? (
                 <img src={previewPlaceholder} alt="bookcover" />
               ) : (
-                <img src={preview} alt="bookcover" />
+                <img src={bookcover} alt="bookcover" />
               )}
             </div>
           )}
@@ -103,8 +107,7 @@ export default function AddBookForm({
             id="bookcover"
             accept=".png, .jpeg, .jpg"
             onChange={event => {
-              onGetBookCoverPreview(event)
-              getPreview(event)
+              getBookcoverPreview(event)
             }}
           />
           <label aria-label="select bookcover" htmlFor="bookcover">
